formatting changes, add checklocks.py back, remove unused imports, simplify input validation on channel/forum lock, add variable  to bot.py, add a few minor TODOs
