+++ src/constants.py
@@ -3,8 +3,12 @@ import os
 TOKEN = os.environ.get("IGCSEBOT_TOKEN")
 LINK = os.environ.get("MONGO_LINK")
 
-GUILD_ID = 576460042774118420 # 1111128710133854289
-LOG_CHANNEL_ID = 1017792876584906782 # 1152467554321842257
+# top: main server, bottom: beta server. comment the unusued one
+GUILD_ID = 576460042774118420
+# GUILD_ID = 1111128710133854289  
+
+LOG_CHANNEL_ID = 1017792876584906782
+# LOG_CHANNEL_ID = 1185411597888393307
 WELCOME_CHANNEL_ID = 930088940654956575
 CREATE_DM_CHANNEL_ID = 895961641219407923
